Moving my laptop version of leapyear so it merges cleanly (or rather doesnt conflict with my repository version of leapyear which is 'finished' but based entirely on Kevlin Henney's instructions and not something I wrote from the ground up.
Moving:
	- nfractions to fractions.py
	  'cause why do I need to prefix it with
n?
